+++ nmrsim/bin/__init__.py
@@ -1,0 +1,2 @@
+# bin may require an __init__ file for importlib.resources. See:
+# https://importlib-resources.readthedocs.io/en/latest/using.html