+++ CHANGELOG.rst
@@ -24,6 +24,16 @@ Working towards a Version 1.0.0 release, the author interprets the terms below a
 * **Version 1.0.0 release**: API is stable. The package is available on PyPI (and perhaps conda).
 
 
+Unreleased
+----------
+Changed
+^^^^^^^
+nmrsim.qp now uses importlib.resources.path to find the nmrsim/bin folder,
+instead of relying on the use of __file__.
+For users, this means that if your application uses nmrsim as a dependency,
+and you want to freeze the application (e.g. with PyInstaller or PyOxidizer),
+the frozen app should now find the bin folder and contents.
+
 0.3.0 - 2019-11-08 (beta release)
 ---------------------------------
 Added